--- rustfmt.toml
@@ -1,4 +1,0 @@
-reorder_imports = false
-reorder_modules = false
-use_field_init_shorthand = true
-use_try_shorthand = true
